Add getUserByHeroId endpoint
The code changes add a new endpoint `getUserByHeroId` to the `Heroes.controller.ts` file. This endpoint retrieves the user associated with a hero by their ID. It returns the user information if found, or returns an error message if either the hero or the user is not found.

This commit message follows the established convention of using a prefix (`feat`) to indicate a new feature or functionality being added.